Moved our .tsx and .scss files to /pages and /styles (REVIEWING THE .scss IS PENDING)
